src/main/java/boj/problems/step6/No2908.java
package boj.problems.step6;

import java.io.BufferedReader;
import java.io.BufferedWriter;
import java.io.IOException;
import java.io.InputStreamReader;
import java.io.OutputStreamWriter;
import java.util.StringTokenizer;

public class No2908 {
    public static void main(String[] args) throws IOException {
        BufferedReader input = new BufferedReader(new InputStreamReader(System.in));
        BufferedWriter output = new BufferedWriter(new OutputStreamWriter(System.out));

        output.write(String.valueOf(solve(input)));

        input.close();
        output.flush();
        output.close();
    }

    static int solve(BufferedReader input) throws IOException {
        StringTokenizer st = new StringTokenizer(input.readLine());

        int a = reverseNumber(st.nextToken());
        int b = reverseNumber(st.nextToken());

        return Math.max(a, b);
    }

    private static int reverseNumber(String str) {
        StringBuilder sb = new StringBuilder(str);

        return Integer.parseInt(sb.reverse().toString());
    }
}
